Create python sub-package for new command.
