+++ turmyx/commands.py
@@ -1,4 +1,3 @@
-from dataclasses import dataclass, field, asdict
 import shutil
 from dataclasses import dataclass, field, asdict
 from subprocess import Popen
